Start to split out Image into its own section
Also, at this point numpy and PIL are not optional.
So remove `_image_requires` which was numpy and PIL.
