+++ mathics/builtin/drawing/image-misc.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+# FIXME - move the rest into builtin.image
 """
 Image[] and image-related functions
 
@@ -11,70 +12,40 @@ module to work.
 # the top level.
 sort_order = "mathics.builtin.image-and-image-related-functions"
 
-import base64
 import functools
 import math
 import os.path as osp
 from collections import defaultdict
-from copy import deepcopy
-from typing import Tuple
 
-from mathics.builtin.base import AtomBuiltin, Builtin, String, Test
-from mathics.builtin.box.image import ImageBox
-from mathics.builtin.colors.color_internals import (
-    colorspaces as known_colorspaces,
-    convert_color,
-)
-from mathics.core.atoms import (
-    Atom,
-    Integer,
-    Integer0,
-    Integer1,
-    MachineReal,
-    Rational,
-    Real,
-)
+import numpy
+import PIL
+
+from mathics.builtin.base import Builtin, String
+from mathics.builtin.colors.color_internals import colorspaces as known_colorspaces
+from mathics.builtin.image.base import Image, _SkimageBuiltin
+from mathics.core.atoms import Integer, Integer0, Integer1, MachineReal, Rational, Real
 from mathics.core.convert.expression import to_mathics_list
 from mathics.core.convert.python import from_python
 from mathics.core.evaluation import Evaluation
 from mathics.core.expression import Expression
 from mathics.core.list import ListExpression
 from mathics.core.symbols import Symbol, SymbolDivide, SymbolNull, SymbolTrue
-from mathics.core.systemsymbols import SymbolImage, SymbolRule
+from mathics.core.systemsymbols import SymbolRule
 from mathics.eval.image import (
     convolve,
     extract_exif,
-    get_image_size_spec,
     matrix_to_numpy,
-    numpy_flip,
     numpy_to_matrix,
     pixels_as_float,
     pixels_as_ubyte,
     pixels_as_uint,
-    resize_width_height,
 )
 
 SymbolColorQuantize = Symbol("ColorQuantize")
-SymbolImage = Symbol("Image")
 SymbolMatrixQ = Symbol("MatrixQ")
 SymbolThreshold = Symbol("Threshold")
 
-# Note a list of packages that are needed for image Builtins.
-_image_requires = ("numpy", "PIL")
-_skimage_requires = _image_requires + ("skimage", "scipy", "matplotlib", "networkx")
-
-try:
-    import warnings
-
-    import numpy
-    import PIL
-    import PIL.ImageEnhance
-    import PIL.ImageFilter
-    import PIL.ImageOps
-
-except ImportError:
-    pass
-
+_skimage_requires = ("skimage", "scipy", "matplotlib", "networkx")
 
 try:
     import skimage.filters
@@ -83,37 +54,15 @@ except ImportError:
 else:
     have_skimage_filters = True
 
-from io import BytesIO
-
 # The following classes are used to allow inclusion of
 # Builtin Functions only when certain Python packages
 # are available. They do this by setting the `requires` class variable.
 
 
-class _ImageBuiltin(Builtin):
-    requires = _image_requires
-
-
-class _ImageTest(Test):
-    """
-    Testing Image Builtins -- those function names ending with "Q" -- that require scikit-image.
-    """
-
-    requires = _image_requires
-
-
-class _SkimageBuiltin(_ImageBuiltin):
-    """
-    Image Builtins that require scikit-image.
-    """
-
-    requires = _skimage_requires
-
-
 # Code related to Mathics Functions that import and export.
 
 
-class ImageExport(_ImageBuiltin):
+class ImageExport(Builtin):
     """
     <dl>
       <dt> 'ImageExport["path", $image$]'
@@ -134,7 +83,7 @@ class ImageExport(_ImageBuiltin):
             return evaluation.message("ImageExport", "noimage")
 
 
-class ImageImport(_ImageBuiltin):
+class ImageImport(Builtin):
     """
     <dl>
       <dt> 'ImageImport["path"]'
@@ -175,7 +124,7 @@ class ImageImport(_ImageBuiltin):
         return ListExpression(*image_list_expression)
 
 
-class _ImageArithmetic(_ImageBuiltin):
+class _ImageArithmetic(Builtin):
     messages = {"bddarg": "Expecting a number, image, or graphics instead of `1`."}
 
     @staticmethod
@@ -320,7 +269,7 @@ class ImageSubtract(_ImageArithmetic):
     summary_text = "build an image substracting pixel values of another image "
 
 
-class RandomImage(_ImageBuiltin):
+class RandomImage(Builtin):
     """
     <url>:WMA link:https://reference.wolfram.com/language/ref/RandomImage.html</url>
 
@@ -392,446 +341,7 @@ class RandomImage(_ImageBuiltin):
         return Image(data, cs)
 
 
-class ImageResize(_ImageBuiltin):
-    """
-    <url>:WMA link:https://reference.wolfram.com/language/ref/ImageResize.html</url>
-
-    <dl>
-      <dt>'ImageResize[$image$, $width$]'
-      <dd>
-
-      <dt>'ImageResize[$image$, {$width$, $height$}]'
-      <dd>
-    </dl>
-
-    The Resampling option can be used to specify how to resample the image. Options are:
-    <ul>
-      <li>Automatic
-      <li>Bicubic
-      <li>Bilinear
-      <li>Box
-      <li>Hamming
-      <li>Lanczos
-      <li>Nearest
-    </ul>
-
-    See <url>
-    :Pillow Filters:
-    https://pillow.readthedocs.io/en/stable/handbook/concepts.html#filters</url>\
-    for a description of these.
-
-    S> alice = Import["ExampleData/MadTeaParty.gif"]
-     = -Image-
-
-    S> shape = ImageDimensions[alice]
-     = {640, 487}
-
-    S> ImageResize[alice, shape / 2]
-     = -Image-
-
-    The default sampling method is "Bicubic" which has pretty good upscaling \
-    and downscaling quality. However "Box" is the fastest:
-
-
-    S> ImageResize[alice, shape / 2, Resampling -> "Box"]
-     = -Image-
-    """
-
-    messages = {
-        "imgrssz": "The size `1` is not a valid image size specification.",
-        "imgrsm": "Invalid resampling method `1`.",
-        "gaussaspect": "Gaussian resampling needs to maintain aspect ratio.",
-        "skimage": "Please install scikit-image to use Resampling -> Gaussian.",
-    }
-
-    options = {"Resampling": "Automatic"}
-    summary_text = "resize an image"
-
-    def eval_resize_width(self, image, s, evaluation, options):
-        "ImageResize[image_Image, s_, OptionsPattern[ImageResize]]"
-        old_w = image.pixels.shape[1]
-        if s.has_form("List", 1):
-            width = s.elements[0]
-        else:
-            width = s
-        w = get_image_size_spec(old_w, width)
-        if w is None:
-            return evaluation.message("ImageResize", "imgrssz", s)
-        if s.has_form("List", 1):
-            height = width
-        else:
-            height = Symbol("Automatic")
-        return self.eval_resize_width_height(image, width, height, evaluation, options)
-
-    def eval_resize_width_height(self, image, width, height, evaluation, options):
-        "ImageResize[image_Image, {width_, height_}, OptionsPattern[ImageResize]]"
-        # resampling method
-        resampling = self.get_option(options, "Resampling", evaluation)
-        if (
-            isinstance(resampling, Symbol)
-            and resampling.get_name() == "System`Automatic"
-        ):
-            resampling_name = "Bicubic"
-        else:
-            resampling_name = resampling.value
-
-        # find new size
-        old_w, old_h = image.pixels.shape[1], image.pixels.shape[0]
-        w = get_image_size_spec(old_w, width)
-        h = get_image_size_spec(old_h, height)
-        if h is None or w is None:
-            return evaluation.message(
-                "ImageResize", "imgrssz", to_mathics_list(width, height)
-            )
-
-        # handle Automatic
-        old_aspect_ratio = old_w / old_h
-        if w == 0 and h == 0:
-            # if both width and height are Automatic then use old values
-            w, h = old_w, old_h
-        elif w == 0:
-            w = max(1, h * old_aspect_ratio)
-        elif h == 0:
-            h = max(1, w / old_aspect_ratio)
-
-        if resampling_name != "Gaussian":
-            # Gaussian need to unrounded values to compute scaling ratios.
-            # round to closest pixel for other methods.
-            h, w = int(round(h)), int(round(w))
-
-        # perform the resize
-        return resize_width_height(image, w, h, resampling_name, evaluation)
-
-
-class ImageReflect(_ImageBuiltin):
-    """
-    <url>
-    :WMA link:
-    https://reference.wolfram.com/language/ref/ImageReflect.html</url>
-    <dl>
-      <dt>'ImageReflect[$image$]'
-      <dd>Flips $image$ top to bottom.
-
-      <dt>'ImageReflect[$image$, $side$]'
-      <dd>Flips $image$ so that $side$ is interchanged with its opposite.
-
-      <dt>'ImageReflect[$image$, $side_1$ -> $side_2$]'
-      <dd>Flips $image$ so that $side_1$ is interchanged with $side_2$.
-    </dl>
-
-    >> ein = Import["ExampleData/Einstein.jpg"];
-    >> ImageReflect[ein]
-     = -Image-
-    >> ImageReflect[ein, Left]
-     = -Image-
-    >> ImageReflect[ein, Left -> Top]
-     = -Image-
-
-    #> ein == ImageReflect[ein, Left -> Left] == ImageReflect[ein, Right -> Right] == ImageReflect[ein, Top -> Top] == ImageReflect[ein, Bottom -> Bottom]
-     = True
-    #> ImageReflect[ein, Left -> Right] == ImageReflect[ein, Right -> Left] == ImageReflect[ein, Left] == ImageReflect[ein, Right]
-     = True
-    #> ImageReflect[ein, Bottom -> Top] == ImageReflect[ein, Top -> Bottom] == ImageReflect[ein, Top] == ImageReflect[ein, Bottom]
-     = True
-    #> ImageReflect[ein, Left -> Top] == ImageReflect[ein, Right -> Bottom]     (* Transpose *)
-     = True
-    #> ImageReflect[ein, Left -> Bottom] == ImageReflect[ein, Right -> Top]     (* Anti-Transpose *)
-     = True
-
-    #> ImageReflect[ein, x -> Top]
-     : x -> Top is not a valid 2D reflection specification.
-     = ImageReflect[-Image-, x -> Top]
-    """
-
-    summary_text = "reflect an image"
-    rules = {
-        "ImageReflect[image_Image]": "ImageReflect[image, Top -> Bottom]",
-        "ImageReflect[image_Image, Top|Bottom]": "ImageReflect[image, Top -> Bottom]",
-        "ImageReflect[image_Image, Left|Right]": "ImageReflect[image, Left -> Right]",
-    }
-
-    messages = {"bdrfl2": "`1` is not a valid 2D reflection specification."}
-
-    def eval(self, image, orig, dest, evaluation: Evaluation):
-        "ImageReflect[image_Image, Rule[orig_, dest_]]"
-        if isinstance(orig, Symbol) and isinstance(dest, Symbol):
-            specs = [orig.get_name(), dest.get_name()]
-            specs.sort()  # `Top -> Bottom` is the same as `Bottom -> Top`
-
-        def anti_transpose(i):
-            return numpy.flipud(numpy.transpose(numpy.flipud(i)))
-
-        def no_op(i):
-            return i
-
-        method = {
-            ("System`Bottom", "System`Top"): numpy.flipud,
-            ("System`Left", "System`Right"): numpy.fliplr,
-            ("System`Left", "System`Top"): numpy.transpose,
-            ("System`Right", "System`Top"): anti_transpose,
-            ("System`Bottom", "System`Left"): anti_transpose,
-            ("System`Bottom", "System`Right"): numpy.transpose,
-            ("System`Bottom", "System`Bottom"): no_op,
-            ("System`Top", "System`Top"): no_op,
-            ("System`Left", "System`Left"): no_op,
-            ("System`Right", "System`Right"): no_op,
-        }.get(tuple(specs), None)
-
-        if method is None:
-            return evaluation.message(
-                "ImageReflect", "bdrfl2", Expression(SymbolRule, orig, dest)
-            )
-
-        return Image(method(image.pixels), image.color_space)
-
-
-class ImageRotate(_ImageBuiltin):
-    """
-
-    <url>:WMA link:https://reference.wolfram.com/language/ref/ImageRotate.html</url>
-
-    <dl>
-    <dt>'ImageRotate[$image$]'
-      <dd>Rotates $image$ 90 degrees counterclockwise.
-    <dt>'ImageRotate[$image$, $theta$]'
-      <dd>Rotates $image$ by a given angle $theta$
-    </dl>
-
-    >> ein = Import["ExampleData/Einstein.jpg"];
-
-    >> ImageRotate[ein]
-     = -Image-
-
-    >> ImageRotate[ein, 45 Degree]
-     = -Image-
-
-    >> ImageRotate[ein, Pi / 4]
-     = -Image-
-
-    #> ImageRotate[ein, ein]
-     : Angle -Image- should be a real number, one of Top, Bottom, Left, Right, or a rule from one to another.
-     = ImageRotate[-Image-, -Image-]
-    """
-
-    messages = {
-        "imgang": "Angle `1` should be a real number, one of Top, Bottom, Left, Right, or a rule from one to another."
-    }
-
-    rules = {"ImageRotate[i_Image]": "ImageRotate[i, 90 Degree]"}
-
-    summary_text = "rotate an image"
-
-    def eval(self, image, angle, evaluation: Evaluation):
-        "ImageRotate[image_Image, angle_]"
-
-        # FIXME: this test I suppose is okay in that it checks more or less what is needed.
-        # However there might be a better test like for Real-valued-ness which could be used
-        # instead.
-        py_angle = (
-            angle.round_to_float(evaluation)
-            if hasattr(angle, "round_to_float")
-            else None
-        )
-
-        if py_angle is None:
-            return evaluation.message("ImageRotate", "imgang", angle)
-
-        def rotate(im):
-            return im.rotate(
-                180 * py_angle / math.pi, resample=PIL.Image.BICUBIC, expand=True
-            )
-
-        return image.filter(rotate)
-
-
-class ImagePartition(_ImageBuiltin):
-    """
-    <url>:WMA link:https://reference.wolfram.com/language/ref/ImagePartition.html</url>
-
-    <dl>
-      <dt>'ImagePartition[$image$, $s$]'
-      <dd>Partitions an image into an array of $s$ x $s$ pixel subimages.
-
-      <dt>'ImagePartition[$image$, {$w$, $h$}]'
-      <dd>Partitions an image into an array of $w$ x $h$ pixel subimages.
-    </dl>
-
-    >> lena = Import["ExampleData/lena.tif"];
-    >> ImageDimensions[lena]
-     = {512, 512}
-    >> ImagePartition[lena, 256]
-     = {{-Image-, -Image-}, {-Image-, -Image-}}
-
-    >> ImagePartition[lena, {512, 128}]
-     = {{-Image-}, {-Image-}, {-Image-}, {-Image-}}
-
-    #> ImagePartition[lena, 257]
-     = {{-Image-}}
-    #> ImagePartition[lena, 512]
-     = {{-Image-}}
-    #> ImagePartition[lena, 513]
-     = {}
-    #> ImagePartition[lena, {256, 300}]
-     = {{-Image-, -Image-}}
-
-    #> ImagePartition[lena, {0, 300}]
-     : {0, 300} is not a valid size specification for image partitions.
-     = ImagePartition[-Image-, {0, 300}]
-    """
-
-    summary_text = "divide an image in an array of sub-images"
-    rules = {"ImagePartition[i_Image, s_Integer]": "ImagePartition[i, {s, s}]"}
-
-    messages = {"arg2": "`1` is not a valid size specification for image partitions."}
-
-    def eval(self, image, w: Integer, h: Integer, evaluation: Evaluation):
-        "ImagePartition[image_Image, {w_Integer, h_Integer}]"
-        py_w = w.value
-        py_h = h.value
-        if py_w <= 0 or py_h <= 0:
-            return evaluation.message("ImagePartition", "arg2", ListExpression(w, h))
-        pixels = image.pixels
-        shape = pixels.shape
-
-        # drop blocks less than w x h
-        parts = []
-        for yi in range(shape[0] // py_h):
-            row = []
-            for xi in range(shape[1] // py_w):
-                p = pixels[yi * py_h : (yi + 1) * py_h, xi * py_w : (xi + 1) * py_w]
-                row.append(Image(p, image.color_space))
-            if row:
-                parts.append(row)
-        return from_python(parts)
-
-
-class ImageAdjust(_ImageBuiltin):
-    """
-
-    <url>:WMA link:
-    https://reference.wolfram.com/language/ref/ImageAdjust.html</url>
-
-    <dl>
-      <dt>'ImageAdjust[$image$]'
-      <dd>adjusts the levels in $image$.
-
-      <dt>'ImageAdjust[$image$, $c$]'
-      <dd>adjusts the contrast in $image$ by $c$.
-
-      <dt>'ImageAdjust[$image$, {$c$, $b$}]'
-      <dd>adjusts the contrast $c$, and brightness $b$ in $image$.
-
-      <dt>'ImageAdjust[$image$, {$c$, $b$, $g$}]'
-      <dd>adjusts the contrast $c$, brightness $b$, and gamma $g$ in $image$.
-    </dl>
-
-    >> lena = Import["ExampleData/lena.tif"];
-    >> ImageAdjust[lena]
-     = -Image-
-    """
-
-    summary_text = "adjust levels, brightness, contrast, gamma, etc"
-    rules = {
-        "ImageAdjust[image_Image, c_?RealNumberQ]": "ImageAdjust[image, {c, 0, 1}]",
-        "ImageAdjust[image_Image, {c_?RealNumberQ, b_?RealNumberQ}]": "ImageAdjust[image, {c, b, 1}]",
-    }
-
-    def eval_auto(self, image, evaluation: Evaluation):
-        "ImageAdjust[image_Image]"
-        pixels = pixels_as_float(image.pixels)
-
-        # channel limits
-        axis = (0, 1)
-        cmaxs, cmins = pixels.max(axis=axis), pixels.min(axis=axis)
-
-        # normalise channels
-        scales = cmaxs - cmins
-        if not scales.shape:
-            scales = numpy.array([scales])
-        scales[scales == 0.0] = 1
-        pixels -= cmins
-        pixels /= scales
-        return Image(pixels, image.color_space)
-
-    def eval_contrast_brightness_gamma(self, image, c, b, g, evaluation: Evaluation):
-        "ImageAdjust[image_Image, {c_?RealNumberQ, b_?RealNumberQ, g_?RealNumberQ}]"
-
-        im = image.pil()
-
-        # gamma
-        g = g.round_to_float()
-        if g != 1:
-            im = PIL.ImageEnhance.Color(im).enhance(g)
-
-        # brightness
-        b = b.round_to_float()
-        if b != 0:
-            im = PIL.ImageEnhance.Brightness(im).enhance(b + 1)
-
-        # contrast
-        c = c.round_to_float()
-        if c != 0:
-            im = PIL.ImageEnhance.Contrast(im).enhance(c + 1)
-
-        return Image(numpy.array(im), image.color_space)
-
-
-class Blur(_ImageBuiltin):
-    """
-    <url>:WMA link:https://reference.wolfram.com/language/ref/Blur.html</url>
-
-    <dl>
-      <dt>'Blur[$image$]'
-      <dd>gives a blurred version of $image$.
-
-      <dt>'Blur[$image$, $r$]'
-      <dd>blurs $image$ with a kernel of size $r$.
-    </dl>
-
-    >> lena = Import["ExampleData/lena.tif"];
-    >> Blur[lena]
-     = -Image-
-    >> Blur[lena, 5]
-     = -Image-
-    """
-
-    summary_text = "blur an image"
-    rules = {
-        "Blur[image_Image]": "Blur[image, 2]",
-        "Blur[image_Image, r_?RealNumberQ]": "ImageConvolve[image, BoxMatrix[r] / Total[Flatten[BoxMatrix[r]]]]",
-    }
-
-
-class Sharpen(_ImageBuiltin):
-    """
-
-    <url>:WMA link:https://reference.wolfram.com/language/ref/Sharpen.html</url>
-
-    <dl>
-      <dt>'Sharpen[$image$]'
-      <dd>gives a sharpened version of $image$.
-
-      <dt>'Sharpen[$image$, $r$]'
-      <dd>sharpens $image$ with a kernel of size $r$.
-    </dl>
-
-    >> lena = Import["ExampleData/lena.tif"];
-    >> Sharpen[lena]
-     = -Image-
-    >> Sharpen[lena, 5]
-     = -Image-
-    """
-
-    summary_text = "sharpen version of an image"
-    rules = {"Sharpen[i_Image]": "Sharpen[i, 2]"}
-
-    def eval(self, image, r, evaluation: Evaluation):
-        "Sharpen[image_Image, r_?RealNumberQ]"
-        f = PIL.ImageFilter.UnsharpMask(r.round_to_float())
-        return image.filter(lambda im: im.filter(f))
-
-
-class GaussianFilter(_ImageBuiltin):
+class GaussianFilter(Builtin):
     """
     <url>:WMA link:https://reference.wolfram.com/language/ref/GaussianFilter.html</url>
 
@@ -860,7 +370,7 @@ class GaussianFilter(_ImageBuiltin):
 # morphological image filters
 
 
-class PillowImageFilter(_ImageBuiltin):
+class PillowImageFilter(Builtin):
     """
 
     ## <url>:PillowImageFilter:</url>
@@ -994,7 +504,7 @@ def _matrix(rows):
     return ListExpression(*[ListExpression(*r) for r in rows])
 
 
-class BoxMatrix(_ImageBuiltin):
+class BoxMatrix(Builtin):
     """
 
     <url>:WMA link:https://reference.wolfram.com/language/ref/BoxMatrix.html</url>
@@ -1017,7 +527,7 @@ class BoxMatrix(_ImageBuiltin):
         return _matrix([[Integer1] * s] * s)
 
 
-class DiskMatrix(_ImageBuiltin):
+class DiskMatrix(Builtin):
     """
     <url>:WMA link:https://reference.wolfram.com/language/ref/DiskMatrix.html</url>
 
@@ -1047,7 +557,7 @@ class DiskMatrix(_ImageBuiltin):
         return _matrix(rows())
 
 
-class DiamondMatrix(_ImageBuiltin):
+class DiamondMatrix(Builtin):
     """
 
     <url>:WMA link:https://reference.wolfram.com/language/ref/DiamondMatrix.html</url>
@@ -1085,7 +595,7 @@ class DiamondMatrix(_ImageBuiltin):
         return _matrix(rows())
 
 
-class ImageConvolve(_ImageBuiltin):
+class ImageConvolve(Builtin):
     """
     <url>:WMA link:https://reference.wolfram.com/language/ref/ImageConvolve.html</url>
 
@@ -1239,7 +749,7 @@ class MorphologicalComponents(_SkimageBuiltin):
 # color space
 
 
-class ImageColorSpace(_ImageBuiltin):
+class ImageColorSpace(Builtin):
     """
     <url>
     :WMA link:
@@ -1262,7 +772,7 @@ class ImageColorSpace(_ImageBuiltin):
         return String(image.color_space)
 
 
-class ColorQuantize(_ImageBuiltin):
+class ColorQuantize(Builtin):
     """
     <url>
     :WMA link:
@@ -1304,7 +814,7 @@ class ColorQuantize(_ImageBuiltin):
         return Image(numpy.array(im), "RGB")
 
 
-class Threshold(_ImageBuiltin):
+class Threshold(Builtin):
     """
 
     <url>:WMA link:https://reference.wolfram.com/language/ref/Threshold.html</url>
@@ -1363,7 +873,7 @@ class Threshold(_ImageBuiltin):
         return MachineReal(float(threshold))
 
 
-class Binarize(_ImageBuiltin):
+class Binarize(Builtin):
     """
     <url>:WMA link:https://reference.wolfram.com/language/ref/Binarize.html</url>
 
@@ -1411,7 +921,7 @@ class Binarize(_ImageBuiltin):
         return Image(mask1 * mask2, "Grayscale")
 
 
-class ColorSeparate(_ImageBuiltin):
+class ColorSeparate(Builtin):
     """
     <url>
     :WMA link:
@@ -1437,7 +947,7 @@ class ColorSeparate(_ImageBuiltin):
         return ListExpression(*images)
 
 
-class ColorCombine(_ImageBuiltin):
+class ColorCombine(Builtin):
     """
     <url>:WMA link:https://reference.wolfram.com/language/ref/ColorCombine.html</url>
 
@@ -1506,7 +1016,7 @@ def _linearize(a):
     return numpy.where(a == h[lower], lower, upper).reshape(orig_shape), n
 
 
-class Colorize(_ImageBuiltin):
+class Colorize(Builtin):
     """
     <url>:WMA link:https://reference.wolfram.com/language/ref/Colorize.html</url>
 
@@ -1574,7 +1084,7 @@ class Colorize(_ImageBuiltin):
 # pixel access
 
 
-class ImageData(_ImageBuiltin):
+class ImageData(Builtin):
     """
 
     <url>:WMA link:
@@ -1625,234 +1135,7 @@ class ImageData(_ImageBuiltin):
         return from_python(numpy_to_matrix(pixels))
 
 
-class ImageTake(_ImageBuiltin):
-    """
-    Crop Image <url>:WMA link:
-    https://reference.wolfram.com/language/ref/ImageTake.html</url>
-    <dl>
-      <dt>'ImageTake[$image$, $n$]'
-      <dd>gives the first $n$ rows of $image$.
-
-      <dt>'ImageTake[$image$, -$n$]'
-      <dd>gives the last $n$ rows of $image$.
-
-      <dt>'ImageTake[$image$, {$r1$, $r2$}]'
-      <dd>gives rows $r1$, ..., $r2$ of $image$.
-
-      <dt>'ImageTake[$image$, {$r1$, $r2$}, {$c1$, $c2$}]'
-      <dd>gives a cropped version of $image$.
-    </dl>
-
-    Crop to the include only the upper half (244 rows) of an image:
-    >> alice = Import["ExampleData/MadTeaParty.gif"]; ImageTake[alice, 244]
-     = -Image-
-
-    Now crop to the include the lower half of that image:
-    >> ImageTake[alice, -244]
-     = -Image-
-
-    Just the text around the hat:
-    >> ImageTake[alice, {40, 150}, {500, 600}]
-     = -Image-
-
-    """
-
-    summary_text = "crop image"
-
-    # FIXME: this probably should be moved out since WMA docs
-    # suggest this kind of thing is done across many kinds of
-    # images.
-    def _image_slice(self, image, i1: Integer, i2: Integer, axis):
-        """
-        Extracts a slice of an image and return a slice
-        indicting a slice, a function flip, that will
-        reverse the pixels in an image if necessary.
-        """
-        n = image.pixels.shape[axis]
-        py_i1 = min(max(i1.value - 1, 0), n - 1)
-        py_i2 = min(max(i2.value - 1, 0), n - 1)
-
-        def flip(pixels):
-            if py_i1 > py_i2:
-                return numpy_flip(pixels, axis)
-            else:
-                return pixels
-
-        return slice(min(py_i1, py_i2), 1 + max(py_i1, py_i2)), flip
-
-    # The reason it is hard to make a rules that turn Image[image, n],
-    # or Image[, {r1, r2} into the generic form Image[image, {r1, r2},
-    # {c1, c2}] there can be negative numbers, e.g. -n. Also, that
-    # missing values, in particular r2 and c2, when filled out can be
-    # dependent on the size of the image.
-
-    # FIXME: create common functions to process ranges.
-    # FIXME: fix up and use _image_slice.
-
-    def eval_n(self, image, n: Integer, evaluation: Evaluation):
-        "ImageTake[image_Image, n_Integer]"
-        py_n = n.value
-        max_y, max_x = image.pixels.shape[:2]
-        if py_n >= 0:
-            adjusted_n = min(py_n, max_y)
-            pixels = image.pixels[:adjusted_n]
-            box_coords = (0, 0, max_x, adjusted_n)
-        elif py_n < 0:
-            adjusted_n = max(0, max_y + py_n)
-            pixels = image.pixels[adjusted_n:]
-            box_coords = (0, adjusted_n, max_x, max_y)
-
-        if hasattr(image, "pillow"):
-            pillow = image.pillow.crop(box_coords)
-            pixels = numpy.asarray(pillow)
-            return Image(pixels, image.color_space, pillow=pillow)
-
-        return Image(pixels, image.color_space, pillow=pillow)
-
-    def eval_rows(self, image, r1: Integer, r2: Integer, evaluation: Evaluation):
-        "ImageTake[image_Image, {r1_Integer, r2_Integer}]"
-
-        first_row = r1.value
-        last_row = r2.value
-
-        max_row, max_col = image.pixels.shape[:2]
-        adjusted_first_row = (
-            min(first_row, max_row) if first_row > 0 else max(0, max_row + first_row)
-        )
-        adjusted_last_row = (
-            min(last_row, max_row) if last_row > 0 else max(0, max_row + first_row)
-        )
-
-        # More complicated in that it reverses the data?
-        # if adjusted_first_row > adjusted_last_row:
-        #     adjusted_first_row, adjusted_last_row = adjusted_last_row, adjusted_first_row
-
-        pixels = image.pixels[adjusted_first_row:adjusted_last_row]
-
-        if hasattr(image, "pillow"):
-            box_coords = (0, adjusted_first_row, max_col, adjusted_last_row)
-            pillow = image.pillow.crop(box_coords)
-            pixels = numpy.asarray(pillow)
-            return Image(pixels, image.color_space, pillow=pillow)
-
-        pixels = image.pixels[adjusted_first_row:adjusted_last_row]
-        return Image(pixels, image.color_space, pillow=pillow)
-
-    def eval_rows_cols(
-        self, image, r1: Integer, r2: Integer, c1: Integer, c2: Integer, evaluation
-    ):
-        "ImageTake[image_Image, {r1_Integer, r2_Integer}, {c1_Integer, c2_Integer}]"
-
-        first_row = r1.value
-        last_row = r2.value
-        first_col = c1.value
-        last_col = c2.value
-
-        max_row, max_col = image.pixels.shape[:2]
-        adjusted_first_row = (
-            min(first_row, max_row) if first_row > 0 else max(0, max_row + first_row)
-        )
-        adjusted_last_row = (
-            min(last_row, max_row) if last_row > 0 else max(0, max_row + last_row)
-        )
-        adjusted_first_col = (
-            min(first_col, max_col) if first_col > 0 else max(0, max_col + first_col)
-        )
-        adjusted_last_col = (
-            min(last_col, max_col) if last_col > 0 else max(0, max_col + last_col)
-        )
-
-        # if adjusted_first_row > adjusted_last_row:
-        #     adjusted_first_row, adjusted_last_row = adjusted_last_row, adjusted_first_row
-
-        # if adjusted_first_col > adjusted_last_col:
-        #     adjusted_first_col, adjusted_last_col = adjusted_last_col, adjusted_first_col
-
-        pixels = image.pixels[
-            adjusted_first_col:adjusted_last_col, adjusted_last_row:adjusted_last_row
-        ]
-
-        if hasattr(image, "pillow"):
-            box_coords = (
-                adjusted_first_col,
-                adjusted_first_row,
-                adjusted_last_col,
-                adjusted_last_row,
-            )
-            pillow = image.pillow.crop(box_coords)
-            pixels = numpy.asarray(pillow)
-            return Image(pixels, image.color_space, pillow=pillow)
-
-        pixels = image.pixels[adjusted_first_row:adjusted_last_row]
-        return Image(pixels, image.color_space, pillow=pillow)
-
-    # Older code we can remove after we condence existing code that looks like this
-    #
-    # def eval_rows(self, image, r1: Integer, r2: Integer, evaluation: Evaluation):
-    #     "ImageTake[image_Image, {r1_Integer, r2_Integer}]"
-    #     s, f = self._slice(image, r1, r2, 0)
-    #     return Image(f(image.pixels[s]), image.color_space)
-
-    # def eval_rows_cols(
-    #     self, image, r1: Integer, r2: Integer, c1: Integer, c2: Integer, evaluation
-    # ):
-    #     "ImageTake[image_Image, {r1_Integer, r2_Integer}, {c1_Integer, c2_Integer}]"
-    #     sr, fr = self._slice(image, r1, r2, 0)
-    #     sc, fc = self._slice(image, c1, c2, 1)
-    #     return Image(fc(fr(image.pixels[sr, sc])), image.color_space)
-
-
-class PixelValue(_ImageBuiltin):
-    """
-    <url>
-    :WMA link:
-    https://reference.wolfram.com/language/ref/PixelValue.html</url>
-
-    <dl>
-      <dt>'PixelValue[$image$, {$x$, $y$}]'
-      <dd>gives the value of the pixel at position {$x$, $y$} in $image$.
-    </dl>
-
-    >> lena = Import["ExampleData/lena.tif"];
-    >> PixelValue[lena, {1, 1}]
-     = {0.321569, 0.0862745, 0.223529}
-    #> {82 / 255, 22 / 255, 57 / 255} // N  (* pixel byte values from bottom left corner *)
-     = {0.321569, 0.0862745, 0.223529}
-
-    #> PixelValue[lena, {0, 1}];
-     : Padding not implemented for PixelValue.
-    #> PixelValue[lena, {512, 1}]
-     = {0.72549, 0.290196, 0.317647}
-    #> PixelValue[lena, {513, 1}];
-     : Padding not implemented for PixelValue.
-    #> PixelValue[lena, {1, 0}];
-     : Padding not implemented for PixelValue.
-    #> PixelValue[lena, {1, 512}]
-     = {0.886275, 0.537255, 0.490196}
-    #> PixelValue[lena, {1, 513}];
-     : Padding not implemented for PixelValue.
-    """
-
-    messages = {"nopad": "Padding not implemented for PixelValue."}
-
-    summary_text = "get pixel value of image at a given position"
-
-    def eval(self, image, x, y, evaluation: Evaluation):
-        "PixelValue[image_Image, {x_?RealNumberQ, y_?RealNumberQ}]"
-        x = int(x.round_to_float())
-        y = int(y.round_to_float())
-        height = image.pixels.shape[0]
-        width = image.pixels.shape[1]
-        if not (1 <= x <= width and 1 <= y <= height):
-            return evaluation.message("PixelValue", "nopad")
-        pixel = pixels_as_float(image.pixels)[height - y, x - 1]
-        if isinstance(pixel, (numpy.ndarray, numpy.generic, list)):
-            return ListExpression(*[MachineReal(float(x)) for x in list(pixel)])
-        else:
-            return MachineReal(float(pixel))
-
-
-class PixelValuePositions(_ImageBuiltin):
+class PixelValuePositions(Builtin):
     """
     <url>:WMA link:https://reference.wolfram.com/language/ref/PixelValuePositions.html</url>
 
@@ -1906,7 +1189,7 @@ class PixelValuePositions(_ImageBuiltin):
 # image attribute queries
 
 
-class ImageDimensions(_ImageBuiltin):
+class ImageDimensions(Builtin):
     """
     <url>
     :WMA link:
@@ -1932,7 +1215,7 @@ class ImageDimensions(_ImageBuiltin):
         return to_mathics_list(*image.dimensions(), elements_conversion_fn=Integer)
 
 
-class ImageAspectRatio(_ImageBuiltin):
+class ImageAspectRatio(Builtin):
     """
     <url>:WMA link:
     https://reference.wolfram.com/language/ref/ImageAspectRatio.html</url>
@@ -1958,7 +1241,7 @@ class ImageAspectRatio(_ImageBuiltin):
         return Expression(SymbolDivide, Integer(dim[1]), Integer(dim[0]))
 
 
-class ImageChannels(_ImageBuiltin):
+class ImageChannels(Builtin):
     """
     <url>:WMA link:
     https://reference.wolfram.com/language/ref/ImageChannels.html</url>
@@ -1983,7 +1266,7 @@ class ImageChannels(_ImageBuiltin):
         return Integer(image.channels())
 
 
-class ImageType(_ImageBuiltin):
+class ImageType(Builtin):
     """
     <url>
     :WMA link:https://reference.wolfram.com/language/ref/ImageType.html</url>
@@ -2012,340 +1295,6 @@ class ImageType(_ImageBuiltin):
         return String(image.storage_type())
 
 
-class BinaryImageQ(_ImageTest):
-    """
-    <url>:WMA link:
-    https://reference.wolfram.com/language/ref/BinaryImageQ.html</url>
-
-    <dl>
-      <dt>'BinaryImageQ[$image]'
-      <dd>returns True if the pixels of $image are binary bit values, and False otherwise.
-    </dl>
-
-    S> img = Import["ExampleData/lena.tif"];
-    S> BinaryImageQ[img]
-     = False
-
-    S> BinaryImageQ[Binarize[img]]
-     = ...
-     : ...
-    """
-
-    summary_text = "test whether pixels in an image are binary bit values"
-
-    def test(self, expr):
-        return isinstance(expr, Image) and expr.storage_type() == "Bit"
-
-
-# Image core classes
-
-
-def _image_pixels(matrix):
-    try:
-        pixels = numpy.array(matrix, dtype="float64")
-    except ValueError:  # irregular array, e.g. {{0, 1}, {0, 1, 1}}
-        return None
-    shape = pixels.shape
-    if len(shape) == 2 or (len(shape) == 3 and shape[2] in (1, 3, 4)):
-        return pixels
-    else:
-        return None
-
-
-class ImageQ(_ImageTest):
-    """
-    <url>:WMA link:https://reference.wolfram.com/language/ref/ImageQ.html</url>
-
-    <dl>
-      <dt>'ImageQ[Image[$pixels]]'
-      <dd>returns True if $pixels has dimensions from which an Image can be constructed, and False otherwise.
-    </dl>
-
-    >> ImageQ[Image[{{0, 1}, {1, 0}}]]
-     = True
-
-    >> ImageQ[Image[{{{0, 0, 0}, {0, 1, 0}}, {{0, 1, 0}, {0, 1, 1}}}]]
-     = True
-
-    >> ImageQ[Image[{{{0, 0, 0}, {0, 1}}, {{0, 1, 0}, {0, 1, 1}}}]]
-     = False
-
-    >> ImageQ[Image[{1, 0, 1}]]
-     = False
-
-    >> ImageQ["abc"]
-     = False
-    """
-
-    summary_text = "test whether is a valid image"
-
-    def test(self, expr):
-        return isinstance(expr, Image)
-
-
-class Image(Atom):
-    class_head_name = "System`Image"
-
-    # FIXME: pixels should be optional if pillow is provided.
-    def __init__(self, pixels, color_space, pillow=None, metadata={}, **kwargs):
-        super(Image, self).__init__(**kwargs)
-
-        if pillow is not None:
-            self.pillow = pillow
-
-        self.pixels = pixels
-
-        if len(pixels.shape) == 2:
-            pixels = pixels.reshape(list(pixels.shape) + [1])
-
-        # FIXME: assigning pixels should be done lazily on demand.
-        # Turn pixels into a property? Include a setter?
-
-        self.pixels = pixels
-
-        self.color_space = color_space
-        self.metadata = metadata
-
-        # Set a value for self.__hash__() once so that every time
-        # it is used this is fast. Note that in contrast to the
-        # cached object key, the hash key needs to be unique across all
-        # Python objects, so we include the class in the
-        # event that different objects have the same Python value
-        self.hash = hash(
-            (
-                SymbolImage,
-                self.pixels.tobytes(),
-                self.color_space,
-                frozenset(self.metadata.items()),
-            )
-        )
-
-    def atom_to_boxes(self, form, evaluation: Evaluation) -> ImageBox:
-        """
-        Converts our internal Image object into a PNG base64-encoded.
-        """
-        pixels = pixels_as_ubyte(self.color_convert("RGB", True).pixels)
-        shape = pixels.shape
-
-        width = shape[1]
-        height = shape[0]
-        scaled_width = width
-        scaled_height = height
-
-        # If the image was created from PIL, use that rather than
-        # reconstruct it from pixels which we can get wrong.
-        # In particular getting color-mapping info right can be
-        # tricky.
-        if hasattr(self, "pillow"):
-            pillow = deepcopy(self.pillow)
-        else:
-            pixels_format = "RGBA" if len(shape) >= 3 and shape[2] == 4 else "RGB"
-            pillow = PIL.Image.fromarray(pixels, pixels_format)
-
-        # if the image is very small, scale it up using nearest neighbour.
-        min_size = 128
-        if width < min_size and height < min_size:
-            scale = min_size / max(width, height)
-            scaled_width = int(scale * width)
-            scaled_height = int(scale * height)
-            pillow = pillow.resize(
-                (scaled_height, scaled_width), resample=PIL.Image.NEAREST
-            )
-
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore")
-
-            stream = BytesIO()
-            pillow.save(stream, format="png")
-            stream.seek(0)
-            contents = stream.read()
-            stream.close()
-
-        encoded = base64.b64encode(contents)
-        encoded = b"data:image/png;base64," + encoded
-
-        return ImageBox(
-            String(encoded.decode("utf-8")),
-            Integer(scaled_width),
-            Integer(scaled_height),
-        )
-
-    # __hash__ is defined so that we can store Number-derived objects
-    # in a set or dictionary.
-    def __hash__(self):
-        return self.hash
-
-    def __str__(self):
-        return "-Image-"
-
-    def color_convert(self, to_color_space, preserve_alpha=True):
-        if to_color_space == self.color_space and preserve_alpha:
-            return self
-        else:
-            pixels = pixels_as_float(self.pixels)
-            converted = convert_color(
-                pixels, self.color_space, to_color_space, preserve_alpha
-            )
-            if converted is None:
-                return None
-            return Image(converted, to_color_space)
-
-    def channels(self):
-        return self.pixels.shape[2]
-
-    def default_format(self, evaluation, form):
-        return "-Image-"
-
-    def dimensions(self) -> Tuple[int, int]:
-        shape = self.pixels.shape
-        return shape[1], shape[0]
-
-    def do_copy(self):
-        return Image(self.pixels, self.color_space, self.metadata)
-
-    def filter(self, f):  # apply PIL filters component-wise
-        pixels = self.pixels
-        n = pixels.shape[2]
-        channels = [
-            f(PIL.Image.fromarray(c, "L")) for c in (pixels[:, :, i] for i in range(n))
-        ]
-        return Image(numpy.dstack(channels), self.color_space)
-
-    def get_sort_key(self, pattern_sort=False) -> tuple:
-        if pattern_sort:
-            # If pattern_sort=True, returns the sort key that matches to an Atom.
-            return super(Image, self).get_sort_key(True)
-        else:
-            # If pattern is False, return a sort_key for the expression `Image[]`,
-            # but with a `2` instead of `1` in the 5th position,
-            # and adding two extra fields: the length in the 5th position,
-            # and a hash in the 6th place.
-            return (1, 3, SymbolImage, len(self.pixels), tuple(), 2, hash(self))
-
-    def grayscale(self):
-        return self.color_convert("Grayscale")
-
-    def pil(self):
-
-        if hasattr(self, "pillow") and self.pillow is not None:
-            return self.pillow
-
-        # see https://pillow.readthedocs.io/en/stable/handbook/concepts.html
-
-        n = self.channels()
-
-        if n == 1:
-            dtype = self.pixels.dtype
-
-            if dtype in (numpy.float32, numpy.float64):
-                pixels = self.pixels.astype(numpy.float32)
-                mode = "F"
-            elif dtype == numpy.uint32:
-                pixels = self.pixels
-                mode = "I"
-            else:
-                pixels = pixels_as_ubyte(self.pixels)
-                mode = "L"
-
-            pixels = pixels.reshape(pixels.shape[:2])
-        elif n == 3:
-            if self.color_space == "LAB":
-                mode = "LAB"
-                pixels = self.pixels
-            elif self.color_space == "HSB":
-                mode = "HSV"
-                pixels = self.pixels
-            elif self.color_space == "RGB":
-                mode = "RGB"
-                pixels = self.pixels
-            else:
-                mode = "RGB"
-                pixels = self.color_convert("RGB").pixels
-
-            pixels = pixels_as_ubyte(pixels)
-        elif n == 4:
-            if self.color_space == "CMYK":
-                mode = "CMYK"
-                pixels = self.pixels
-            elif self.color_space == "RGB":
-                mode = "RGBA"
-                pixels = self.pixels
-            else:
-                mode = "RGBA"
-                pixels = self.color_convert("RGB").pixels
-
-            pixels = pixels_as_ubyte(pixels)
-        else:
-            raise NotImplementedError
-
-        return PIL.Image.fromarray(pixels, mode)
-
-    def options(self):
-        return ListExpression(
-            Expression(SymbolRule, String("ColorSpace"), String(self.color_space)),
-            Expression(SymbolRule, String("MetaInformation"), self.metadata),
-        )
-
-    def sameQ(self, other) -> bool:
-        """Mathics SameQ"""
-        if not isinstance(other, Image):
-            return False
-        if self.color_space != other.color_space or self.metadata != other.metadata:
-            return False
-        return numpy.array_equal(self.pixels, other.pixels)
-
-    def storage_type(self):
-        dtype = self.pixels.dtype
-        if dtype in (numpy.float32, numpy.float64):
-            return "Real"
-        elif dtype == numpy.uint32:
-            return "Bit32"
-        elif dtype == numpy.uint16:
-            return "Bit16"
-        elif dtype == numpy.uint8:
-            return "Byte"
-        elif dtype == bool:
-            return "Bit"
-        else:
-            return str(dtype)
-
-    def to_python(self, *args, **kwargs):
-        return self.pixels
-
-
-class ImageAtom(AtomBuiltin):
-    """
-    <url>
-    :WMA link:
-    https://reference.wolfram.com/language/ref/ImageAtom.html</url>
-
-    <dl>
-      <dt>'Image[...]'
-      <dd> produces the internal representation of an image from an array \
-          of values for the pixels.
-    </dl>
-
-    #> Image[{{{1,1,0},{0,1,1}}, {{1,0,1},{1,1,0}}}]
-     = -Image-
-
-    #> Image[{{{0,0,0,0.25},{0,0,0,0.5}}, {{0,0,0,0.5},{0,0,0,0.75}}}]
-     = -Image-
-    """
-
-    summary_text = "get internal representation of an image"
-    requires = _image_requires
-
-    def eval_create(self, array, evaluation: Evaluation):
-        "Image[array_]"
-        pixels = _image_pixels(array.to_python())
-        if pixels is not None:
-            shape = pixels.shape
-            is_rgb = len(shape) == 3 and shape[2] in (3, 4)
-            return Image(pixels.clip(0, 1), "RGB" if is_rgb else "Grayscale")
-        else:
-            return Expression(SymbolImage, array)
-
-
 # complex operations
 
 
@@ -2370,7 +1319,7 @@ class TextRecognize(Builtin):
 
     options = {"Language": '"English"'}
 
-    requires = _image_requires + ("pyocr",)
+    requires = "pyocr"
 
     summary_text = "recognize text in an image"
 
@@ -2465,7 +1414,7 @@ class WordCloud(Builtin):
         "MaxItems": "Automatic",
     }
 
-    requires = _image_requires + ("wordcloud",)
+    requires = ("wordcloud",)
 
     summary_text = "show a word cloud from a list of words"
