+++ mathics/builtin/image/__init__.py
@@ -1,0 +1,3 @@
+"""
+Image Manipulation
+"""
