+++ mathics/builtin/options.py
@@ -12,7 +12,7 @@ https://reference.wolfram.com/language/guide/OptionsManagement.html</url>
 """
 
 from mathics.builtin.base import Builtin, Test, get_option
-from mathics.builtin.drawing.image import Image
+from mathics.builtin.image.base import Image
 from mathics.core.atoms import String
 from mathics.core.evaluation import Evaluation
 from mathics.core.expression import Expression, SymbolDefault, get_default_value
