+++ mathics/eval/image.py
@@ -234,7 +234,7 @@ def resize_width_height(
     """
     workhorse part of ImageResize[] after mathic options have been processed.
     """
-    from mathics.builtin.drawing.image import Image
+    from mathics.builtin.image.base import Image
 
     if resampling_name not in resampling_names2PIL.keys():
         return evaluation.message("ImageResize", "imgrsm", resampling_name)
